+++ test/file/delimited_test.rb
@@ -21,6 +21,19 @@ module NdrImport
         end
       end
 
+      test 'should read streamed csv correctly' do
+        file_path = @permanent_test_files.join('normal.csv')
+        file_object = ::File.open(file_path, 'r')
+        handler = NdrImport::File::Delimited.new(file_object, 'csv', 'col_sep' => nil)
+        handler.tables.each do |tablename, sheet|
+          assert_nil tablename
+          sheet = sheet.to_a
+          assert_equal(('A'..'Z').to_a, sheet[0])
+          assert_equal ['1'] * 26, sheet[1]
+          assert_equal ['2'] * 26, sheet[2]
+        end
+      end
+
       test 'should read pipe correctly' do
         file_path = @permanent_test_files.join('normal_pipe.csv')
         handler = NdrImport::File::Delimited.new(file_path, 'delimited', 'col_sep' => '|')