+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-* No unreleased changes
+### Added
+* Ability to process CSV files as an IO stream
 
 ## 10.1.3 / 2022-12-08
 ### Added